### 2023-08-09 #5748 @editorialbot
[PRE REVIEW]: Dnaapler: A tool to reorient circular microbial genomes
**Submitting author:** @gbouras13 (<a href="http://orcid.org/0000-0002-5885-4186">George Bouras</a>)
**Repository:** https://github.com/gbouras13/dnaapler
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
