### 2023-08-09 #5748 @gbouras13
Some people who would be good reviewers:

- Michael Hall https://reviewers.joss.theoj.org/reviewers/2619 - while he is in the acknowledgements as we copied & modified some of his code, this was done without any of his input & he has had nothing to do with Dnaapler.
- Curtis Kapsak https://reviewers.joss.theoj.org/reviewers/1647
- Justine Debelius https://reviewers.joss.theoj.org/reviewers/2113
- Vini Salazar https://reviewers.joss.theoj.org/reviewers/1974
- Andrea Telatin https://reviewers.joss.theoj.org/reviewers/1383
