### 2023-08-15 5b84e88
test for all command
