### 2023-08-15 669c3c9
save full contig header (record.description) not just id
