### 2023-08-15 c218702
slight fix for BLAST output matching need short contig
