+++ src/dnaapler/utils/all.py
@@ -1,0 +1,262 @@
+import os
+import shutil
+from pathlib import Path
+
+import pandas as pd
+from Bio import SeqIO
+from loguru import logger
+
+from dnaapler.utils.constants import DNAAPLER_DB
+from dnaapler.utils.external_tools import ExternalTool
+from dnaapler.utils.processing import reorient_single_record_bulk
+from dnaapler.utils.validation import validate_custom_db_fasta
+
+
+def all_process_blast_output_and_reorient(input, blast_file, output, prefix):
+    """Processes the blast output, reorients and saves all contigs into os.path.join(output, f"{prefix}_all_reoriented.fasta")
+
+    :param input: input file
+    :param blast_file: blast output file
+    :param output: output directory
+    :param prefix: prefix
+    :return:
+    """
+
+    # define colnames
+    col_list = [
+        "qseqid",
+        "qlen",
+        "sseqid",
+        "slen",
+        "length",
+        "qstart",
+        "qend",
+        "sstart",
+        "send",
+        "pident",
+        "nident",
+        "gaps",
+        "mismatch",
+        "evalue",
+        "bitscore",
+        "qseq",
+        "sseq",
+    ]
+
+    # read in the dataframe from BLAST
+    try:
+        blast_df = pd.read_csv(
+            blast_file, delimiter="\t", index_col=False, names=col_list
+        )
+
+    except Exception:
+        logger.error("There was an issue with parsing the BLAST output file.")
+
+    if isinstance(blast_df, pd.DataFrame) and blast_df.empty:
+        logger.error(
+            "There were 0 BLAST hits. Please check your input file or try dnaapler custom. If you have assembled an understudied species, this may also be the cause."
+        )
+
+    # Initialize the list to store the IDs
+    contigs = []
+    genes = []  # for the gene it reoriented it by
+    starts = []
+    strands = []
+    top_hits = []
+    top_hit_lengths = []
+    covered_lens = []
+    coverages = []
+    idents = []
+    identitys = []
+
+    reoriented_output_file = os.path.join(output, f"{prefix}_all_reoriented.fasta")
+
+    # Read the FASTA file and extract the IDs
+    for record in SeqIO.parse(input, "fasta"):
+        contig = record.description
+        # need this to match for BLAST
+        short_contig = record.id
+        print(contig)
+
+        contigs.append(contig)
+
+        # Filter the DataFrame where 'qseqid' matches 'contig'
+        filtered_df = blast_df[blast_df["qseqid"] == short_contig]
+
+        length_of_df = len(filtered_df)
+
+        # no hits at all then just write to the file
+        if length_of_df == 0:
+            with open(reoriented_output_file, "a") as out_fa:
+                SeqIO.write(record, out_fa, "fasta")
+
+            # no hit save for the output DF
+            message = "No_BLAST_hits"
+            genes.append(message)
+            starts.append(message)
+            strands.append(message)
+            top_hits.append(message)
+            top_hit_lengths.append(message)
+            covered_lens.append(message)
+            coverages.append(message)
+            idents.append(message)
+            identitys.append(message)
+
+        else: # there is at least one BLAST hit
+
+            # determine the numbers of repA, dnaA and terL
+
+            # UniRef90 is in string for repA
+            # phrog is in string for terL
+            # DNAA is in string for dnaA
+
+
+            # counts
+
+
+            counts = {
+            'dnaA': filtered_df[filtered_df['sseqid'].str.contains('DNAA', case=False)].shape[0],
+            'terL': filtered_df[filtered_df['sseqid'].str.contains('phrog', case=False)].shape[0],
[…]
