+++ src/dnaapler/__init__.py
@@ -27,6 +27,7 @@ from dnaapler.utils.validation import (
     validate_custom_db_fasta,
     validate_fasta,
     validate_fasta_bulk,
+    validate_ignore_file
 )
 
 from dnaapler.utils.all import (all_process_blast_output_and_reorient)
@@ -624,8 +625,6 @@ def bulk(
     end_dnaapler(start_time)
 
 
-
-
 """
 all subcommand
 """
@@ -642,6 +641,13 @@ all subcommand
     help="e value for blastx",
     show_default=True,
 )
+@click.option(
+    "--ignore",
+    default="",
+    help="TSV file listing contigs (one per row) that are to be ignored",
+    type=click.Path(),
+    show_default=False
+)
 def all(
     ctx,
     input,
@@ -650,6 +656,7 @@ def all(
     prefix,
     evalue,
     force,
+    ignore,
     **kwargs,
 ):
     """Reorients multiple contigs to begin with any of dnaA, repA or terL"""
@@ -669,13 +676,32 @@ def all(
     # validate e value
     check_evalue(evalue)
 
+    # checks if the ignore file exists and contains text
+    
+    if ignore != "":
+        logger.info(f"You have specified contigs to ignore in {ignore}.")
+        exists_contains_txt = validate_ignore_file(ignore) 
+
     # runs bulk BLAST
     run_bulk_blast(ctx, input, output, prefix, gene, evalue, threads, custom_db = None)
 
     # rerorients blast
     blast_file = os.path.join(output, f"{prefix}_blast_output.txt")
 
-    all_process_blast_output_and_reorient(input, blast_file, output, prefix)
+
+    ### ignore
+
+    if exists_contains_txt is False:
+        logger.warning(f"{ignore} contains no text. No contigs will be ignored")
+        ignore_list = []
+    else:
+    # gets all contigs in the ignore
+    # will split by space so short_contig only (to match BLAST)
+        with open(ignore) as f:
+            ignore_dict = {x.rstrip().split()[0] for x in f}
+        ignore_list = list(ignore_dict)
+
+    all_process_blast_output_and_reorient(input, blast_file, output, prefix, ignore_list)
 
     # end dnaapler
     end_dnaapler(start_time)
