+++ docs/run.md
@@ -11,7 +11,7 @@ You can use BLAST with multiple threads using the `-t` or `--threads` parameters
 
 `dnaapler` will not overwrite an output directory if it already exists by default. To force overwrite, please use `-f` or `--force`.
 
-Finally, for the BLAST based subcommands (`chromosome`, `phage`, `plasmid` or `custom`), if no BLAST hit is found, by default `dnaapler` will error and exit. 
+Finally, for the BLAST based subcommands (`chromosome`, `phage`, `plasmid`, `custom` or `all`), if no BLAST hit is found, by default `dnaapler` will error and exit. 
 
 However, you can decide to autocomplete `dnaapler` using the `-a` or `--autocomplete` parameters along with `mystery` or `nearest`, which will then run those subcommands to reorient your sequence.
 
@@ -226,4 +226,48 @@ Options:
                          custom blast database to reorient your sequence
                          however you want. Must be specified if -m custom is
                          specified.
+```
+
+### all
+
+
+`dnaapler all` is designed to simultaneously orient multiple contigs that can be a mix of chromosomes, plasmids and phages.
+
+If a contig has BLAST hits for both dnaA and terL or repA, dnaA will be chosen for reorientation.
+
+If a contig has BLAST hits for both terL and repA (but not dnaA), repA will be chosen for reorientation.
+
+You can also specify a text file with `--ignore` that lists all contigs (based on their header) to be ignored during reorientation.
+
+e.g. the file (`ignored_contigs.txt`) needs to be formatted as follows:
+
+```
+contig_1
+contig_2
+```
+
+Your input FASTA must also have at least 2 contigs.
+
+Example usage to reorient a number of contigs in `input.fasta`, ignoring all contigs with headers denoted in  `ignored_contigs.txt`
+
+```
+dnaapler all -i input.fasta -o output_directory_path -t 8  --ignore ignored_contigs.txt
+```
+
+```
+Usage: dnaapler all [OPTIONS]
+
+  Reorients multiple contigs to begin with any of dnaA, repA or terL
+
+Options:
+  -h, --help             Show this message and exit.
+  -V, --version          Show the version and exit.
+  -i, --input PATH       Path to input file in FASTA format  [required]
+  -o, --output PATH      Output directory   [default: output.dnaapler]
+  -t, --threads INTEGER  Number of threads to use with BLAST  [default: 1]
+  -p, --prefix TEXT      Prefix for output files  [default: dnaapler]
+  -f, --force            Force overwrites the output directory
+  -e, --evalue TEXT      e value for blastx  [default: 1e-10]
+  --ignore PATH          TSV file listing contigs (one per row) that are to be
+                         ignored
 ```
