### 2023-08-15 7c674f1
add reoriented fasta to all
