+++ build/meta.yaml
@@ -1,29 +1,0 @@
-package:
-  name: "dnaapler"
-  version: "0.0.1"
-
-build:
-  number: 0
-  noarch: generic
-
-source:
-  git_url: https://github.com/gbouras13/dnaapler
-
-requirements:
- run:
-  - blast
-  - biopython
-  - python >=3
-  - pandas
-
-
-
-about:
-  home: https://github.com/gbouras13/dnaapler
-  license: MIT
-  license_familY: MIT
-  summary: "Orients complete bacterial whole genome chromosome assemblies to begin with dnaa gene"
-
-extra:
-  recipe-maintainers:
-    - gbouras13
