### 2023-08-25 9d73df5
fix typo in logging
Just one minor typo
