### 2023-08-29 bf0abdb
Fix #38
Fix a little bug in the code that led to crashes in reorienting assemblies after the BLAST step
