### 2023-08-30 ae911de
bump version and all test to cover #38
