### 2023-09-04 #5748 @Kevin-Mattheus-Moerman
@gbouras13 thanks for this submission. I am the AEiC for this track and here to help with initial steps. Apologies for the delay so far. I have been out of the office for a while but am picking this up now.

Note I have just added the `query-scope` flag. This is to trigger a scope review by the editorial board. We usually do this for submissions that are rather small in size. The scope review will determine if this work is judged to conform to our [substantial scholarly effort criteria](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This review should take about 2 weeks to complete.
