+++ pyproject.toml
@@ -22,7 +22,7 @@ loguru = ">=0.5.3"
 pyyaml = ">=6.0"
 pandas = ">=1.4.2"
 biopython = ">=1.76"
-pyrodigal = ">=3.0.0,<4.0.0"
+pyrodigal = ">=3.0.0"
 
 
 [tool.poetry.dev-dependencies]
