### 2023-10-20 #5748 @fboehm
hi, @gbouras13 - my name is Fred, and I'm a JOSS editor. I'll work with you throughout the review of your submission.
The first step is for us to find two reviewers. Thank you so much for suggesting reviewers above. I'll contact them and others with similar expertise to ask about their interest and availability to review.
It's probably easiest for me if we communicate here (ie, in this github thread), although I'm happy to communicate about any sensitive matters via email.
Once we identify two reviewers, we'll "officially" start the review. Subsequent notifications and communications will then occur in another (not yet created) github thread. You should get notifications (ie, github "@" mentions for your username) when all of these events occur.
Please let me know if you have any questions. if not, then it should be a matter of a few days before we get the reviewers and start the review.
Thanks!
