### 2023-10-21 #5748 @fboehm
Thank you, @gbouras13 for the updated information and transparency.

It looks like we've found two reviewers - I'll officially add them as reviewers in a moment. We can then start the reviews.

Thank you, @mkerin and @vinisalazar ! I'll add you as reviewers in a moment. Please follow the hyperlink that gets printed in the comment thread here to find the "review issue" for this submission. (PS - where we're now communicating is termed by JOSS as the "pre-review" thread). We'll communicate in the "review issue" / thread starting right away.
