### 2023-10-21 #5968 @editorialbot
[REVIEW]: Dnaapler: A tool to reorient circular microbial genomes
**Submitting author:** @gbouras13 (<a href="http://orcid.org/0000-0002-5885-4186">George Bouras</a>)
**Repository:** https://github.com/gbouras13/dnaapler
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.0
