### 2023-10-25 ef5b9fa
Fixes #44 to use the most overlapping BLAST tophit CDS if no valid start codon found
