+++ src/dnaapler/utils/processing.py
@@ -1,11 +1,11 @@
 import os
+from pathlib import Path
 
 import pandas as pd
-from Bio import SeqIO
-from loguru import logger
-from pathlib import Path
 import pyrodigal
+from Bio import SeqIO
 from Bio.SeqRecord import SeqRecord
+from loguru import logger
 
 
 def process_blast_output_and_reorient(input, blast_file, out_file, gene: str):
@@ -82,44 +82,41 @@ def process_blast_output_and_reorient(input, blast_file, out_file, gene: str):
     # https://github.com/gbouras13/dnaapler/issues/44
 
     else:
-        if blast_df["qseq"][0][0] in ["M", "V", "L"] and (
-            blast_df["sstart"][0] == 1
-        ):
+        if blast_df["qseq"][0][0] in ["M", "V", "L"] and (blast_df["sstart"][0] == 1):
             reorient_sequence(blast_df, input, out_file, gene, overlapping_orf=False)
-        else: # this will reorient the sequence with the orf that overlaps the tophit by the most
+        else:  # this will reorient the sequence with the orf that overlaps the tophit by the most
             # warn if the top hit doesnt begin with a valid start codon
             logger.warning(
-                    f"The top {gene} blastx hit did not have a BLAST alignment beginning with a valid start codon M, V or L."
-                )
+                f"The top {gene} blastx hit did not have a BLAST alignment beginning with a valid start codon M, V or L."
+            )
             logger.warning(
-                    "The CDS with the most overlap with this blastx hit according to Pyrodigal will instead be used to reorient the genome."
-                )
+                "The CDS with the most overlap with this blastx hit according to Pyrodigal will instead be used to reorient the genome."
+            )
             reorient_sequence(blast_df, input, out_file, gene, overlapping_orf=True)
-               
-
 
         blast_success = True
 
     return blast_success
 
 
-def reorient_sequence(blast_df: pd.DataFrame, input: Path, 
-                      out_file: Path, 
-                      gene: str, 
-                      overlapping_orf: bool) -> None:
-    
+def reorient_sequence(
+    blast_df: pd.DataFrame,
+    input: Path,
+    out_file: Path,
+    gene: str,
+    overlapping_orf: bool,
+) -> None:
     # get the start
-    
+
     ident = blast_df["nident"][0]
     top_hit = blast_df["sseqid"][0]
     top_hit_length = blast_df["slen"][0]
     covered_len = blast_df["length"][0]
     identity = round(ident / covered_len * 100, 2)
     coverage = round(covered_len / top_hit_length * 100, 2)
-    
+
     # top blast hit starts at 1
     if overlapping_orf is False:
-
         # basic stats
         dnaa_start_on_chromosome = blast_df["qstart"][0]
         strand = "forward"
@@ -135,7 +132,9 @@ def reorient_sequence(blast_df: pd.DataFrame, input: Path,
         logger.info(
             f"{covered_len} AAs were covered by the best hit, with an overall coverage of {coverage}%."
         )
-        logger.info(f"{ident} AAs were identical, with an overall identity of {identity}%.")
+        logger.info(
+            f"{ident} AAs were identical, with an overall identity of {identity}%."
+        )
         logger.info("Re-orienting.")
 
         ####################
@@ -167,7 +166,6 @@ def reorient_sequence(blast_df: pd.DataFrame, input: Path,
 
     # top blastx hit does not start at 1
     else:
-
         # get start and end of tophit
         start_tophit = blast_df["qstart"][0]
         end_tophit = blast_df["qend"][0]
@@ -177,13 +175,15 @@ def reorient_sequence(blast_df: pd.DataFrame, input: Path,
 
         # there will only be 1 record
[…]
