### 2023-10-25 c9192cc
add autocomplete to dnaapler all
