+++ src/dnaapler/__init__.py
@@ -10,7 +10,12 @@ from loguru import logger
 
 from dnaapler.utils.all import all_process_blast_output_and_reorient
 from dnaapler.utils.bulk import bulk_process_blast_output_and_reorient, run_bulk_blast
-from dnaapler.utils.cds_methods import run_blast_based_method, run_mystery, run_nearest
+from dnaapler.utils.cds_methods import (
+    run_blast_based_method,
+    run_largest,
+    run_mystery,
+    run_nearest,
+)
 from dnaapler.utils.constants import DNAAPLER_DB
 from dnaapler.utils.external_tools import ExternalTool
 from dnaapler.utils.util import (
@@ -88,6 +93,32 @@ def common_options(func):
     return func
 
 
+def autocomplete_options(func):
+    """
+    autocomplete options
+    """
+    options = [
+        click.option(
+            "-a",
+            "--autocomplete",
+            type=click.STRING,
+            callback=validate_choice_autocomplete,
+            default="none",
+            help="Choose an option to autocomplete reorientation if BLAST based approach fails.\nMust be one of: none, mystery, largest, or nearest [default: none]",
+        ),
+        click.option(
+            "--seed_value",
+            help="Random seed to ensure reproducibility.",
+            type=int,
+            default=13,
+            show_default=True,
+        ),
+    ]
+    for option in reversed(options):
+        func = option(func)
+    return func
+
+
 @click.group()
 @click.help_option("--help", "-h")
 @click.version_option(get_version(), "--version", "-V")
@@ -112,21 +143,7 @@ Chromosome command
     help="e value for blastx",
     show_default=True,
 )
-@click.option(
-    "-a",
-    "--autocomplete",
-    type=click.STRING,
-    callback=validate_choice_autocomplete,
-    default="none",
-    help="Choose an option to autocomplete reorientation if BLAST based approach fails.\nMust be one of: none, mystery or nearest [default: none]",
-)
-@click.option(
-    "--seed_value",
-    help="Random seed to ensure reproducibility.",
-    type=int,
-    default=13,
-    show_default=True,
-)
+@autocomplete_options
 def chromosome(
     ctx,
     input,
@@ -190,21 +207,7 @@ Plasmid command
     help="e value for blastx",
     show_default=True,
 )
-@click.option(
-    "-a",
-    "--autocomplete",
-    type=click.STRING,
-    callback=validate_choice_autocomplete,
-    default="none",
-    help="Choose an option to autocomplete reorientation if BLAST based approach fails.\nbe one of: none, mystery or nearest [default: none]",
-)
-@click.option(
-    "--seed_value",
-    help="Random seed to ensure reproducibility.",
-    type=int,
-    default=13,
-    show_default=True,
-)
+@autocomplete_options
 def plasmid(
     ctx,
     input,
@@ -268,21 +271,7 @@ Phage command
     help="e value for blastx",
     show_default=True,
 )
-@click.option(
-    "-a",
-    "--autocomplete",
-    type=click.STRING,
-    callback=validate_choice_autocomplete,
-    default="none",
-    help="Choose an option to autocomplete reorientation if BLAST based approach fails.\nMust be one of: none, mystery or nearest [default: none]",
-)
-@click.option(
-    "--seed_value",
-    help="Random seed to ensure reproducibility.",
-    type=int,
-    default=13,
-    show_default=True,
-)
+@autocomplete_options
 def phage(
     ctx,
     input,
@@ -353,21 +342,7 @@ custom command
     type=click.Path(),
     required=True,
 )
-@click.option(
-    "-a",
-    "--autocomplete",
-    type=click.STRING,
[…]
