### 2023-10-25 5219e9e
add typing and dupe header checks
