+++ src/dnaapler/utils/cds_methods.py
@@ -14,7 +14,25 @@ from dnaapler.utils.processing import (
 )
 
 
-def run_mystery(ctx, input, seed_value, output, prefix):
+def run_mystery(ctx, input: Path, seed_value: int, output: Path, prefix: str) -> None:
+    """
+    Reorients a DNA sequence with a random Coding DNA Sequence (CDS) from the input file.
+
+    Args:
+        ctx (object): A context or environment object, possibly from a command-line tool.
+        input (str): Path to the input DNA sequence file in FASTA format.
+        seed_value (int): Seed value for randomization.
+        output (str): Path to the output directory where the reoriented sequence will be saved.
+        prefix (str): Prefix for the output file name.
+
+    Returns:
+        None
+
+    This function searches for CDS using Pyrodigal in the input DNA sequence, then randomly selects
+    one CDS (not the first or last) and reorients the DNA sequence with respect to that CDS.
+    It saves the reoriented sequence in the specified output directory with the given prefix.
+    """
+
     # get number of records of input
     logger.info("Searching for CDS with pyrodigal")
     orf_finder = pyrodigal.GeneFinder(meta=True)
@@ -56,7 +74,23 @@ def run_mystery(ctx, input, seed_value, output, prefix):
         reorient_sequence_random(input, output_processed_file, start, strand)
 
 
-def run_nearest(ctx, input, output, prefix):
+def run_nearest(ctx, input: Path, output: Path, prefix: str) -> None:
+    """
+    Reorients a DNA sequence to begin with the first Coding DNA Sequence (CDS) found in the input file.
+
+    Args:
+        ctx (object): A context or environment object, possibly from a command-line tool.
+        input (str): Path to the input DNA sequence file in FASTA format.
+        output (str): Path to the output directory where the reoriented sequence will be saved.
+        prefix (str): Prefix for the output file name.
+
+    Returns:
+        None
+
+    This function searches for CDS using Pyrodigal in the input DNA sequence, then reorients the DNA
+    sequence to begin with the first CDS found. The reoriented sequence is saved in the specified
+    output directory with the given prefix.
+    """
     # get number of records of input
     logger.info("Searching for CDS with pyrodigal")
     orf_finder = pyrodigal.GeneFinder(meta=True)
@@ -93,13 +127,27 @@ def run_nearest(ctx, input, output, prefix):
         reorient_sequence_random(input, output_processed_file, start, strand)
 
 
-def run_largest(ctx, input, seed_value, output, prefix):
+def run_largest(ctx, input: Path, output: Path, prefix: str) -> None:
+    """
+    Reorients a DNA sequence to begin with the largest Coding DNA Sequence (CDS) found in the input file.
+
+    Args:
+        ctx (object): A context or environment object, possibly from a command-line tool.
+        input (Path): Path to the input DNA sequence file in FASTA format.
+        seed_value (int): Seed value for randomization.
+        output (Path): Path to the output directory where the reoriented sequence will be saved.
+        prefix (str): Prefix for the output file name.
+
+    Returns:
+        None
+
+    This function searches for CDS using Pyrodigal in the input DNA sequence, then reorients the DNA
+    sequence to begin with the largest CDS found based on the size of the coding region. The reoriented
+    sequence is saved in the specified output directory with the given prefix.
+    """
     # get number of records of input
     logger.info("Searching for CDS with pyrodigal")
     orf_finder = pyrodigal.GeneFinder(meta=True)
-
-    # set seed
[…]
