### 2023-11-01 gbouras13/dnaapler#47 @vinisalazar
[JOSS Review] - Text corrections and suggestions
Hi,

attached are some corrections and suggestions for the manuscript text based on my review.

https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/paper/paper.md?plain=1#L36

**Suggestions**

* Line 15: Remove "advanced" preceding "technologies", or replace with a more descriptive term such as "high-throughput".

https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/paper/paper.md?plain=1#L38

**Corrections**:
* [x] Line 19: Plasmids should be referred as something other than a "microorganism", such as a "biological entity" or some other term. As the first line reads, it sounds as bacteriophages and viruses are different entities, should be rewritten as something such as "bacteriophages and other viruses".

**Suggestions**

* Line 19: Archaea and Bacteria should be capitalised.
* Line 20, 26: The second and third sentences both begin in the same way, with "however" and "therefore". The second sentence can be rewritten as "Once assembled, a circular genome is represented [...]".

https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/paper/paper.md?plain=1#L46

**Suggestions**
* Line 40: add a comma after the word "difficult".
* Line 41: replace "Python 3" with "Python". Swap "can easily be" for "can be easily". Could be rewritten as: "command-line tool written in Python that performs the reorientation of microbial genomes, and can be easily integrated [...]".
* Line 43:  Rewrite "uses the Basic Local Alignment [...]" for "uses BLAST".
* Line 44: replace "only external dependency" with "only mandatory external dependency". **Edit: I initially assumed Pyrodigal was also an external dependency, but now I've realised it is a Python package. However, wouldn't Pyrodigal depend on the Prodigal tool, and thus Prodigal would be an optional external dependency? Please confirm this, thank you.**

**Corrections**
* [x] Line 43-44: explicitly state the purpose of BLAST and Pyrodigal, i.e. "uses BLAST or Pyrodigal as input files/for alignment/some other purpose".

Overall I believe that this paragraph must be more descriptive in **what exactly does Dnaapler do**, as this is only implied through context.

https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/paper/paper.md?plain=1#L61

**Corrections**
* [x] Line 60-61: this should be moved to the "Availability" section.

https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/paper/paper.md?plain=1#L67

**Suggestions**
* Line 65: replace "Github" with "GitHub" for consistency with Line 64.

---

Edit: ~~Apologies, I accidentally submitted this issue before finishing writing it, so I will be editing it before it's complete.~~ This issue is now complete.

cc openjournals/joss-reviews#5968
