### 2023-11-01 gbouras13/dnaapler#48 @vinisalazar
[JOSS Review] - Documentation
Hi,

attached are some suggestions for the documentation based on my review:

https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/README.md?plain=1#L20

* Include `-c bioconda`, like in the command specified in Line 95 of the same file. This should prevent many first-time users from running into a "package not found" error, as it is unlikely they will have read the [Beginner Conda Installation](https://dnaapler.readthedocs.io/en/latest/install/) section of the docs. You may also link this section here.

https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/README.md?plain=1#L106

* Specify that this is only required if installing with pip (would Prodigal be required too if the user wants to use the Pyrodigal commands?)

https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/docs/install.md?plain=1#L46

* This list numbering is rendering incorrectly in the readthedocs page.

---

I will update this issue if I find anything else.

cc openjournals/joss-reviews#5968
