### 2023-11-01 #5968 @vinisalazar
The authors present a tool for automated reorientation of microbial genomes. The implementation is simple and elegant, with high code quality, and employs best practices for scientific software.

This is a high quality submission and I wholly recommend it for publication in JOSS.

Having this said, my only concern with the manuscript (and the accompanying software documentation) is the lack of actual example usages of the software, either on artificial or real-world data. Although the authors demonstrate how to utilise the different subcommands, there is no explicit demonstration of what the software does, that is, showing an output representing a reoriented genome. The testing suite in the software package contains example data that can presumably be used to demonstrate this. Although I don't think that this should be a blocker to accepting the manuscript, I believe that the manuscript and/or the documentation would be greatly enriched by this addition. One suggestion would be to show figures comparing "no treatment" vs "reoriented" genomes on a genome browser such as Geneious.

Some additional points:
- I would suggest renaming the `mystery` subcommand as `random` as that is more descriptive of what it does.
- I understand that Dnaapler reorients genomes by placing specific marker genes at the start. But how does it decide the order of the remaining genes, does it follow the order that they were in the original FASTA file? Is there any guarantee that this is the actual order of CDSs in the genome?
