### 2023-11-02 gbouras13/dnaapler#49 @mkerin
[JOSS Review] Typos in CONTRIBUTING.md
Re [joss-reviews/issues/5968](https://github.com/openjournals/joss-reviews/issues/5968)

Hi @gbouras13,

There are a couple of places in the CONTRIBUTING.md file which reference one of your other projects (`phrokka`) rather than `dnaapler`. Could you please take a look and make sure that the contribution guidelines are appropriate for this repo? It might also be nice to link them in the main README.md.
