### 2023-11-02 gbouras13/dnaapler#50 @mkerin
[JOSS Review] Authorship list
Re [joss-reviews/issues/5968](https://github.com/openjournals/joss-reviews/issues/5968)

Hi @gbouras13

I have a question about the list of authors in the JOSS paper. I can see that 3/5 of the authors have made contributions to the repo, but I am unclear about the contributions of the remaining authors. Could you please clarify what their contribution was?

I note that I've seen some JOSS papers provide a statement of author contributions according to the [Credit](https://beta.elsevier.com/researcher/author/policies-and-guidelines/credit-author-statement) guidelines, but this doesn't appear to be required.
