### 2023-11-02 gbouras13/dnaapler#51 @mkerin
[JOSS Review] Installation instructions in the readthedocs
Re [joss-reviews/issues/5968](https://github.com/openjournals/joss-reviews/issues/5968)

The [installation instructions](https://dnaapler.readthedocs.io/en/latest/install/) for dnaapler currently provide a choice between two channels:
1. Installing via conda
2. Installing via pip
and then there is a 'Beginner Conda Installation' subsection.

The following things jump out to me:
A. I find it confusing that if one follows the 'beginner conda installation' instructions, then one ends up following a different codepath to (1) and installing with mamba instead.
B. Installing mamba via conda is [no longer recommended](https://mamba.readthedocs.io/en/latest/mamba-installation.html#mamba-install).
C. I'm not sure if pip & conda can coexist peacefully in the same virtual environment. Does it make sense to suggest to your users allow users to install dnaapler via pip if BLAST is only available in bioconda?

To resolve (A) & (B) you could:
- make installation instructions for all 3 channels (conda, mamba, pip) available at the same level.
- refer users to the right section of the conda/mamba website to help them install these package managers rather than hardcoding the commands yourself.
