### 2023-11-02 gbouras13/dnaapler#47 @mkerin
In addition, the citation for `(Mallawaarachchi et al., 2023)` points to a bioRxiv paper, but this paper has since been published at
https://academic.oup.com/bioinformatics/article/39/10/btad586/7280146
Could you update the citation please?
