### 2023-11-03 gbouras13/dnaapler#52 @mkerin
[JOSS Review] Clarifying the fallback behaviour from BLAST to pyrodigal
Re openjournals/joss-reviews#5968

Hello,

I have tried to common sense check the output of `dnaapler` on the provided test data by running the following
```
dnaapler all -i tests/test_data/SAOMS1.fasta -o output.all -f
```
In case it is helpful I have attached a .zip folder of the output that I get from this command.

Two things are surprising to me:
1. We get logs from running the BLAST external dependency, but only the log for the standard error is populated (`output.all/logs/blastx_7d5612cc02ea76c324567a7b1bfeb1dfe9d64fbfc4b176959385cdc60514d626.err`)
2. The primary log states that `top blastx hit ... not begin with a valid start codon` and dnaapler falls back to reorientating the sequence using pyrodigal

Could you explain why BLAST is producing errors in (1), and then the BLAST ouptut is being ignored in (2)? I worry that (1) implies that BLAST has hard-crashed, and then the error has been swallowed because `dnaapler` has fallen back to using `pyrodigal`.

[output.all.zip](https://github.com/gbouras13/dnaapler/files/13251150/output.all.zip)
