### 2023-11-03 #5968 @mkerin
Hi all - I have now finished my review.

I think dnaapler is a strong submission to JOSS. The library has through documentation, appears to be well tested, and I can see how separating out the re-orientation of bacterial genomes into a separate library is a valuable contribution to the community.

I have left 5 boxes in the checklist unticked, which represent areas where I have outstanding questions / comments:
- For the *authorship*/*community guidelines*/*references*, I have opened issues in the dnaapler repo which I believe are minor only & should be straight forward to resolve.
- On testing the *functionality* of `dnaapler`, I performed basic checks using the test `.fasta` file to verify that `dnaapler` conserves the order of the dna sequence during reorientation. For example, if `transMystery()` and `transAll()` are functions that transform genomes accordinging to `dnaapler mystery` and `dnaapler all` respectively, then `transAll(fasta) == transAll(transMystery(fasta))`. However the logs suggested that `dnaapler` was not actually using BLAST as expected - so it would be great in @gbouras13 could clarify if this is expected (see issue in dnaapler repo for more details).
- On the *example usage*, I agree with @vinisalazar that it would be useful to see examples using real world data & accompanying visualizations of the original & reorientated sequences - if this is not too burdensome.

Once these points are resolved then I will be happy to recommended this paper for publication in JOSS.
