### 2023-11-06 gbouras13/dnaapler#52 @gbouras13
Re https://github.com/openjournals/joss-reviews/issues/5968

The run worked as expected for the test phage.

To explain more fully:

Point 1.

Everytime blastx is run in dnaapler, it uses the class ExternalTool class defined in external_tools.py (https://github.com/gbouras13/dnaapler/blob/main/src/dnaapler/utils/external_tools.py). This is taken (with some modifications) from tbpore https://github.com/mbhall88/tbpore/blob/main/tbpore/external_tools.py.

For any external tool (in dnaapler, this will only be blastx), this will automatically write stdout to the `.out` file if it is not redirected otherwise by the tool, and the stderr to the `.err` file.

Because blastx specifies an output file using `-out`, all stdout is redirected to that file and so the `out` log file is empty. Therefore only the `.err` file is populated. In what you have attached, the blastx output file is `dnaapler_blast_output.txt` and is correct. This is quirk of blastx - what is in that file is the stderr (which is just the command line).

In terms of your concern about blastx crashing, the ExternalTool class has built in exception handling. If a subprocess error is detected, then it will trigger a system exit and tell the user of the error (in red text as well). Given the FASTA validation checks, blastx installation checks and unit tests, I anticipate this will be rare, but if it occurs, this error will be handled.

Point 2.

Prior to v0.4.0, dnaapler would look for a blastx hit alignment from the output file that started with a valid stop codon "M", "V", or "L". If it could not find a start codon, it would notify the user that a hit was found but without a valid start codon and exit, returning the input FASTA as the output.

However, following this issue #44, some more testing by one of the co-authors on her data and a realisation this would exclude many sequences with distant homology to the databases (we anticipated this would be a particular problem for novel phages and plasmids where near perfect matches would be less likely to be in the databases), in v0.4.0 the logic has been changed. (see here for the current implementation https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/src/dnaapler/utils/all.py#L238C50-L238C63 and https://github.com/gbouras13/dnaapler/blob/7c67b3beb930c9e2a1e031612841bc7541aa6915/src/dnaapler/utils/processing.py#L292 )

Now, dnaapler will still look at the top blastx hit as previously. If the alignment does begin with a valid start codon, dnaapler proceeds as usual.

If the alignment doesn't begin with a valid start codon, then instead of warning and exiting, pyrodigal is run to predict all the CDS on the contig.

Dnaapler then calculates the CDS that has the most overlap with the blastx top hit, and chooses that CDS to reorient the sequence with. This should be the CDS most likely to be the terL/repA/dnaA.

In the attached case, that is what happened with the test phage - it chose a CDS starting at 4183 on the forward strand as terL, and choose that to reorient the phage with.
