### 2023-11-06 f23c8a9
clean up contributing file
