+++ CONTRIBUTING.md
@@ -15,7 +15,7 @@ Types of Contributions
 Report Bugs
 ~~~~~~~~~~~
 
-Report bugs at https://github.com/gbouras13/phrokka/issues.
+Report bugs at https://github.com/gbouras13/dnaapler/issues.
 
 If you are reporting a bug, please include:
 
@@ -38,14 +38,14 @@ and "help wanted" is open to whoever wants to implement it.
 Write Documentation
 ~~~~~~~~~~~~~~~~~~~
 
-phrokka could always use more documentation, whether as part of the
-official phrokka docs, in docstrings, or even on the web in blog posts,
+dnaapler could always use more documentation, whether as part of the
+official dnaapler docs, in docstrings, or even on the web in blog posts,
 articles, and such.
 
 Submit Feedback
 ~~~~~~~~~~~~~~~
 
-The best way to send feedback is to file an issue at https://github.com/gbouras13/phrokka/issues.
+The best way to send feedback is to file an issue at https://github.com/gbouras13/dnaapler/issues.
 
 If you are proposing a feature:
 
@@ -57,18 +57,17 @@ If you are proposing a feature:
 Get Started!
 ------------
 
-Ready to contribute? Here's how to set up `phrokka` for local development.
+Ready to contribute? Here's how to set up `dnaapler` for local development.
 
-1. Fork the `phrokka` repo on GitHub.
+1. Fork the `dnaapler` repo on GitHub.
 2. Clone your fork locally::
 
-    $ git clone git@github.com:your_name_here/phrokka.git
+    $ git clone git@github.com:gbouras13/dnaapler.git
 
 3. Install your local copy into a virtualenv. Assuming you have virtualenvwrapper installed, this is how you set up your fork for local development::
 
-    $ mkvirtualenv phrokka
-    $ cd phrokka/
-    $ python setup.py develop
+    $ mkvirtualenv dnaapler
+    $ cd dnaapler/
 
 4. Create a branch for local development::
 
@@ -76,12 +75,12 @@ Ready to contribute? Here's how to set up `phrokka` for local development.
 
    Now you can make your changes locally.
 
-5. When you're done making changes, check that your changes pass flake8 and the
-   tests, including testing other Python versions with tox::
+5. When you're done making changes, check that your changes pass black and isort and the
+   tests::
 
-    $ flake8 phrokka tests
-    $ python setup.py test or pytest
-    $ tox
+    $ black . 
+    $ isort . 
+    $ poetry run pytest tests/
 
    To get flake8 and tox, just pip install them into your virtualenv.
 
@@ -101,28 +100,6 @@ Before you submit a pull request, check that it meets these guidelines:
 1. The pull request should include tests.
 2. If the pull request adds functionality, the docs should be updated. Put
    your new functionality into a function with a docstring, and add the
-   feature to the list in README.rst.
-3. The pull request should work for Python 3.5, 3.6, 3.7 and 3.8, and for PyPy. Check
-   https://travis-ci.com/gbouras13/phrokka/pull_requests
-   and make sure that the tests pass for all supported Python versions.
-
-Tips
-----
-
-To run a subset of tests::
-
-$ pytest tests.test_phrokka
+   feature to the list in README.md and HISTORY.md.
 
 
-Deploying
----------
-
-A reminder for the maintainers on how to deploy.
-Make sure all your changes are committed (including an entry in HISTORY.rst).
-Then run::
-
-$ bump2version patch # possible: major / minor / patch
-$ git push
-$ git push --tags
-
-Travis will then deploy to PyPI if tests pass.
