### 2023-11-06 c73c8fc
clean up contributing file headers
