### 2023-11-06 gbouras13/dnaapler#49 @gbouras13
Thanks @mkerin ,

This has been addressed in #54 and #55. I will also included a link to the README.md in a later PR.
