### 2023-11-06 d5d91bf
add contributing to readme [skip ci]
