### 2023-11-06 b10ab75
add contributing to readme fix typo [skip ci]
