### 2023-11-06 gbouras13/dnaapler#50 @gbouras13
Hi @mkerin ,

I'm not sure how to include this (or if it needs to be) for the manuscript but I will outline here:

GB: conceptualisation, methodology, software, validation, writing - original draft, writing - editing and review

SRG: validation, writing - editing and review

BP: methodology, validation, writing - editing and review

VM: software, writing - editing and review, visualisation, supervision

MJR: software, writing - editing and review, supervision
