### 2023-11-06 c391203
clarify BLAST install with pip [skip ci]
