+++ README.md
@@ -17,7 +17,7 @@ Dnaapler is a simple tool that reorients complete circular microbial genomes.
 
 ```
 # creates conda environment with dnaapler
-conda create -n dnaapler_env dnaapler
+conda create -n dnaapler_env -c bioconda dnaapler
 
 # activates conda environment
 conda activate dnaapler_env
@@ -86,7 +86,7 @@ The full documentation for `dnaapler` can be found [here](https://dnaapler.readt
 
 `dnaapler` requires only BLAST v2.10 or higher as an external dependency. 
 
-Installation from conda is recommended as this will install BLAST automatically.
+Installation from conda is highly recommended as this will install BLAST automatically.
 
 ### Conda
 
@@ -104,13 +104,14 @@ You can also install `dnaapler` with pip.
 pip install dnaapler
 ```
 
-You will need to install BLAST v2.10 or higher separately.
+* Note: If you want to install `dnaapler` with pip, you will need to install BLAST v2.10 or higher separately.
 
 e.g.
 
 ```
 conda install -c bioconda blast>2.9
 ```
+
 
 ## Usage
 
