### 2023-11-06 gbouras13/dnaapler#48 @gbouras13
Hi @vinisalazar ,

Thanks for these. Also linking to #51 regarding your third point as they will be dealt with together.

Your first point has been fixed in the READ.me to make it easier for beginners to follow - namely, by making an empty conda env and then installing dnaapler (I did not get good results trying to do it all in one line, therefore a beginner will have a harder time!).

The second point also has been fixed.

To clarify, prodigal is not required dnaapler at all. This is because pyrodigal is a self contained Python package that reimplements and provides python bindings to Prodigal and therefore does not require it as a dependency. pyrodigal will always be installed with a pip install of dnaapler.

Regarding your third point, install.md is overhauled after #51 - the correct numbering will be used.

George
