### 2023-11-06 gbouras13/dnaapler#51 @gbouras13
Hi @mkerin ,

Thanks for this, all suggestions are implemented.

The only comment I have is regarding your pip and BLAST concern. Usually pip plays fine within conda. And BLAST is available in a lot of installation methods, not just conda.

Realistically, the pip installation option will be for users who do not like using conda or do not have it installed, and are therefore likely to be competent command line users. Therefore, reference to installing conda from BLAST has been removed in place of a comment suggesting BLAST needs to be available in the PATH.

George
