### 2023-11-06 8aefc07
clean up conda install [skip ci]
