### 2023-11-06 716580b
remove ordering and add citation deps [skip ci]
