### 2023-11-06 ecaf95a
add install outpout [skip ci]
