### 2023-11-06 #5968 @gbouras13
Regarding the visualisation/example data question, we think this is feasible to include in the manuscript. We will include an example chromosome and phage visualised by the plot output options in Bakta and Pharokka respectively (with some manual annotation to make the dnaA/terL obvious - it will be clearer for phage due to the genome size).

 We will also outline  tests and examples explicitly in the documentation using data in the tests directory.
