### 2023-11-06 c315ac5
add plots to paper
