+++ paper/paper.bib
@@ -138,7 +138,6 @@
     pages = {btad586},
     year = {2023},
     month = {09},
-    abstract = "{Microbial communities have a profound impact on both human health and various environments. Viruses infecting bacteria, known as bacteriophages or phages, play a key role in modulating bacterial communities within environments. High-quality phage genome sequences are essential for advancing our understanding of phage biology, enabling comparative genomics studies and developing phage-based diagnostic tools. Most available viral identification tools consider individual sequences to determine whether they are of viral origin. As a result of challenges in viral assembly, fragmentation of genomes can occur, and existing tools may recover incomplete genome fragments. Therefore, the identification and characterization of novel phage genomes remain a challenge, leading to the need of improved approaches for phage genome recovery.We introduce Phables, a new computational method to resolve phage genomes from fragmented viral metagenome assemblies. Phables identifies phage-like components in the assembly graph, models each component as a flow network, and uses graph algorithms and flow decomposition techniques to identify genomic paths. Experimental results of viral metagenomic samples obtained from different environments show that Phables recovers on average over 49\\% more high-quality phage genomes compared to existing viral identification tools. Furthermore, Phables can resolve variant phage genomes with over 99\\% average nucleotide identity, a distinction that existing tools are unable to make.Phables is available on GitHub at https://github.com/Vini2/phables.}",
     issn = {1367-4811},
     doi = {10.1093/bioinformatics/btad586},
     url = {https://doi.org/10.1093/bioinformatics/btad586},
@@ -146,7 +145,7 @@
 }
 
 
-@article{Bouras:2023,
+@article{Bouras1:2023,
     author = {Bouras, George and Sheppard, Anna E and Mallawaarachchi, Vijini and Vreugde, Sarah},
     title = "{Plassembler: an automated bacterial plasmid assembly tool}",
     journal = {Bioinformatics},
@@ -155,7 +154,6 @@
     pages = {btad409},
     year = {2023},
     month = {06},
-    abstract = "{With recent advances in sequencing technologies, it is now possible to obtain near-perfect complete bacterial chromosome assemblies cheaply and efficiently by combining a long-read-first assembly approach with short-read polishing. However, existing methods for assembling bacterial plasmids from long-read-first assemblies often misassemble or even miss bacterial plasmids entirely and accordingly require manual curation. Plassembler was developed to provide a tool that automatically assembles and outputs bacterial plasmids using a hybrid assembly approach. It achieves increased accuracy and computational efficiency compared to the existing gold standard tool Unicycler by removing chromosomal reads from the input read sets using a mapping approach.Plassembler is implemented in Python and is installable as a bioconda package using ‘conda install -c bioconda plassembler’. The source code is available on GitHub at https://github.com/gbouras13/plassembler. The full benchmarking pipeline can be found at https://github.com/gbouras13/plassembler\_simulation\_benchmarking, while the benchmarking input FASTQ and output files can be found at https://doi.org/10.5281/zenodo.7996690.}",
     issn = {1367-4811},
     doi = {10.1093/bioinformatics/btad409},
     url = {https://doi.org/10.1093/bioinformatics/btad409},
@@ -169,10 +167,47 @@
 	year = {2023},
 	doi = {10.1101/2023.03.28.534496},
 	publisher = {Cold Spring Harbor Laboratory},
[…]
