### 2023-11-06 #5968 @gbouras13
I also note that `dnaapler all` and `dnaapler largest` have been added since v0.2.0 (submission in August), so I have added them to the same - especially dnaapler all which I anticipate is the most useful command.
