+++ paper/paper.bib
@@ -172,7 +172,6 @@
 	journal = {bioRxiv}
 }
 
-
 @article{Bouras2:2023,
     author = {Bouras, George and Nepal, Roshan and Houtak, Ghais and Psaltis, Alkis James and Wormald, Peter-John and Vreugde, Sarah},
     title = "{Pharokka: a fast scalable bacteriophage annotation tool}",
