+++ paper/paper.md
@@ -50,19 +50,25 @@ Circlator [@Hunt:2015] is the most commonly used dedicated tool for reorienting 
 | chromosome       | Custom database downloaded from Swissprot                                     | dnaA chromosomal replication initiator gene |
 | plasmid          | repA database curated from Unicycler [@Wick:2017]                             | repA plasmid replication initiation gene    |
 | phage            | Prokaryotic Virus Remote Homologous Groups database (PHROGs) [@Terzian:2021]  | terL large terminase subunit gene           |
+| all              | Chromosome, plasmid and phage databases combined                              | dnaA, repA and terL                         |
 | custom           | User specified                                                                | Custom gene                                 |
 | mystery          | Pyrodigal predicted coding sequences                                          | Random CDS                                  |
 | nearest          | Pyrodigal predicted coding sequences                                          | First CDS (nearest to the start)            |
+| largest          | Pyrodigal predicted coding sequences                                          | Largest CDS                                 |
 | bulk             | Either chromosome, plasmid, phage or custom. Requires multiple input contigs. | dnaA, repA, terL or a custom gene           |
 
 
-Specifically, Dnaapler 'chromosome', 'phage' and 'plasmid' subcommands use blastx (protein databases are searched using a translated nucleotide query) to search for the dnaA, terL or repA gene respectively in the input genomes, using built-in amino acid databases for each gene. Dnaapler will check that the first amino acid of the identified start site begins with Methionine, Valine, or Leucine (the 3 most used gene start codons in bacteria and bacteriophages) and will then reorient the genome to begin with this gene in the forward direction. If the 'custom' subcommand is selected, the same process will be conducted but with a user specified amino acid FASTA formatted input database. If the 'mystery' or 'nearest' subcommands are selected, Pyrodigal will be used to predict all coding sequences, and the genome will be reoriented to begin with either a random (mystery) or the first (nearest) CDS, respectively. Dnaapler returns an output directory containing a log file and the genome reoriented as a FASTA formatted file. Finally, the 'bulk' subcommand can be used to reorient multiple input contigs (in a mulitFASTA format file) using either the chromosome, plasmid, phage or custom reorientation strategies.
+Specifically, Dnaapler 'chromosome', 'phage' and 'plasmid' subcommands use blastx (protein databases are searched using a translated nucleotide query) to search for the dnaA, terL or repA gene respectively in the input genomes, using built-in amino acid databases for each gene. Dnaapler 'all' will run a blastx search against all three databases simultaneously, prioritising dnaA hits then repA and finally terL if multiple genes have hits. 
 
-Examples of Dnaapler's functionality on the C333 _Staphylococcus aureus_ chromosome and the C333 Sa3int prophage (the chromosome can be found as accession GCA_030288915.1, sample number SAMN32360890 from Bioproject PRJNA914892 from [@Houtak:2023]) are shown below using the plotting functionalities of Bakta [@Schwengers:2021] and Pharokka [Bouras2:2023].
+Taking the top blastx hit, Dnaapler will check that the first amino acid of the BLAST alignment begins with Methionine, Valine, or Leucine (the 3 most used gene start codons in bacteria and bacteriophages). If it does, then it will then reorient the genome to begin at that position in the forward direction. If it does not, then Pyrodigal will be used to predict all coding sequences. Dnaapler will calculate the CDS with the most overlap to the top blastx hit, and reorient the genome to begin with the start codon of that CDS in the forward direction. 
 
[…]
