### 2023-11-06 7242979
explain the exact coordinate
