### 2023-11-06 gbouras13/dnaapler#47 @gbouras13
@mkerin ,

That has been done thanks.

@vinisalazar , I have made all of the minor changes requested. Regarding the larger points:

1. Prodigal is not a dependency (as answered in #48 ).
2. I have updated the explanation of what Dnaapler is doing in the paper to more accurately to reflect the response to #52.
