### 2023-11-06 gbouras13/dnaapler#49 @mkerin
Thank you for taking a look at this - I agree that the issue is resolved. Closing now.
