### 2023-11-06 gbouras13/dnaapler#51 @mkerin
Hi George. Okay, I'm happy with this. Thank you for the quick response.
