### 2023-11-06 gbouras13/dnaapler#52 @mkerin
Thank you @gbouras13 . In response to the comments above:
1. Okay, I understand why the `.out` file is empty. But why is `blastx` producing any `stderr` output at all? If you run the same blastx command by hand (interative terminal outside of the ExternalTool class) does it produce stderr output?
2. Okay, I can see that there are cases when we want to fall back gracefully. Just to check - does the test suite include a case where blastx returns a valid hit and dnaapler doesn't have to fall back to prodigal?
