### 2023-11-07 gbouras13/dnaapler#52 @gbouras13
Hi @mkerin ,

1. I have looked into this and if you run blastx just in the terminal ( e.g. with 2> blast.err) it captures no stderr. This prompted me to look deeper, and I realised this was caused by a little bug, namely this print statement:

```
print(f"Command line: {self.command_as_str}", file=stderr_fh)
```

https://github.com/gbouras13/dnaapler/blob/637429ce50ca59e007f23b78947e754506a0c4f8/src/dnaapler/utils/external_tools.py#L49

When I remove that print statement, it results in a blank .stderr. Thanks for picking this up! I've made a commit to address this and remove the statement.

2. Yes there are tests for both scenarios - see unit tests here specifically to parse blast output and check that this is being captured

 https://github.com/gbouras13/dnaapler/blob/637429ce50ca59e007f23b78947e754506a0c4f8/tests/test_dnaapler.py#L75

and end-to-end tests here(all of them running on chromosome.fasta will cover the scenario where there is no fallback)

 https://github.com/gbouras13/dnaapler/blob/637429ce50ca59e007f23b78947e754506a0c4f8/tests/test_overall.py#L64

George
