### 2023-11-07 gbouras13/dnaapler#52 @gbouras13
I accidentally changed check_call to call in the first commit - I reverted in the second (sorry for the confusion)
