### 2023-11-07 94a296d
strand forward and back for all when backup overlap pyrodigal
