+++ docs/example.md
@@ -1,0 +1,105 @@
+
+# `dnaapler` examples
+
+You can try running `dnaapler` yourself using test data found in the `tests/test_data/overall_inputs`. These examples assumed you have cloned the dnaapler repository from GitHub and have moved into the directory e.g. and have `dnaapler` [installed](install.md) :
+
+```
+git clone "https://github.com/gbouras13/dnaapler.git"
+cd dnaapler
+```
+
+## Chromosome
+
+This chromosome is from _Staphylococcus aureus_ isolate C333 taken from [Houtak et al](https://www.biorxiv.org/content/10.1101/2023.03.28.534496v1), GenBank accession GCA_030288915.1, Sample Number SAMN32360890 from BioProject PRJNA914892 .
+
+To run `dnaapler chromosome` to reorient the C333 chromosome to begin with the dnaA gene 
+
+```
+dnaapler chromosome -i tests/test_data/overall_inputs/chromosome.fasta -o C333_dnaapler -t 8 -p C333
+```
+
+The output should look like:
+
+```
+2023-11-07 12:08:51.243 | INFO     | dnaapler.utils.validation:instantiate_dirs:23 - Checking the output directory C333_dnaapler
+2023-11-07 12:08:51.251 | INFO     | dnaapler.utils.util:begin_dnaapler:71 - You are using dnaapler version 0.4.0
+2023-11-07 12:08:51.252 | INFO     | dnaapler.utils.util:begin_dnaapler:72 - Repository homepage is https://github.com/gbouras13/dnaapler
+2023-11-07 12:08:51.252 | INFO     | dnaapler.utils.util:begin_dnaapler:73 - Written by George Bouras: george.bouras@adelaide.edu.au
+2023-11-07 12:08:51.252 | INFO     | dnaapler.utils.util:begin_dnaapler:74 - Your input FASTA is tests/test_data/overall_inputs/chromosome.fasta
+2023-11-07 12:08:51.252 | INFO     | dnaapler.utils.util:begin_dnaapler:75 - Your output directory  is C333_dnaapler
+2023-11-07 12:08:51.252 | INFO     | dnaapler.utils.util:begin_dnaapler:76 - You have specified 8 threads to use with blastx
+2023-11-07 12:08:51.252 | INFO     | dnaapler.utils.util:begin_dnaapler:77 - You have specified dnaA gene(s) to reorient your sequence
+2023-11-07 12:08:51.252 | INFO     | dnaapler.utils.util:check_blast_version:115 - Checking BLAST installation.
+2023-11-07 12:08:51.290 | INFO     | dnaapler.utils.util:check_blast_version:135 - BLAST version found is v2.14.1.
+2023-11-07 12:08:51.290 | INFO     | dnaapler.utils.util:check_blast_version:145 - BLAST version is ok.
+2023-11-07 12:08:51.290 | INFO     | dnaapler.utils.util:check_pyrodigal_version:90 - Checking pyrodigal installation.
+2023-11-07 12:08:51.290 | INFO     | dnaapler.utils.util:check_pyrodigal_version:101 - Pyrodigal version is v3.1.1
+2023-11-07 12:08:51.290 | INFO     | dnaapler.utils.util:check_pyrodigal_version:102 - Pyrodigal version is ok.
+2023-11-07 12:08:51.290 | INFO     | dnaapler:chromosome:170 - You have chosen none method to reorient your sequence if the BLAST based method fails.
+2023-11-07 12:08:51.290 | INFO     | dnaapler.utils.validation:validate_fasta:46 - Checking that the input file tests/test_data/overall_inputs/chromosome.fasta is in FASTA format and has only 1 entry.
[…]
