### 2023-11-07 #5968 @gbouras13
Hi @vinisalazar and @mkerin ,

We have added examples of chromosome and phage reorientation using examples from the testing suite (that are from a real study as indicated) and visualisations to the documentation (https://dnaapler.readthedocs.io/en/latest/example/). The two figures have also been added to the paper.

George
