### 2023-11-07 #5968 @gbouras13
@vinisalazar to respond to your comments:

1. I do not want to change the command line interface by renaming mystery to random in case users have implemented these changes in their pipelines (I have personally), though I agree with your point at it being more descriptive. I do like the mystery name as well!
2. With your second point, dnaapler will reorient the genomes, keeping the CDS ordering as they were in the input. You can see this more clearly in the visualisations that have been added (especially the phage one as that is easier to see and also because it was reoriented on the forward strand). A core assumption of dnaapler is that the input sequence(s) should be circular, but this is for the user to make sure of (and it is in the title of the manuscript). In terms of a guarantee that is the order they are in in the genome, that is more of an issue with the assembly - dnaapler requires complete replicons as input that theoretically should be correctly assembled.
3. I have beefed up both the citation section in the documentation and also in the program itself to add references to Prodigal (I am citing this as suggested by the author of Pyrodigal even though Prodigal itself isn't a dependency) , BLAST, and Pyrodigal (https://github.com/gbouras13/dnaapler/commit/716580bb968ccd4e8a8d7aa445e2ddd8e8d57faa) thank you for this.
