### 2023-11-07 bc312a2
beef up docs [skip ci]
