### 2023-11-08 gbouras13/dnaapler#48 @mkerin
This is a quick comment to say that I think there is an unfinished sentence on line 54 of `docs/example.md` here
https://github.com/gbouras13/dnaapler/blob/bf86601136951079218f258452f5ff40733d50f9/docs/example.md?plain=1#L54

I also think that, other than this typo, the new docs/example.md section is excellent.
