### 2023-11-08 gbouras13/dnaapler#52 @mkerin
Hi @gbouras13,

Thank you for the fix! I am happy that this has been resolved - closing the issue now.
