### 2023-11-08 gbouras13/dnaapler#52 @mkerin
Sorry - I have one more question on this subject.

Just for my own edification, why do you only consider the top blast hit before falling back to pyrodigal. Is it possible that blast returns >1 hit, where the second hit has a valid start codon?
