### 2023-11-08 gbouras13/dnaapler#52 @gbouras13
Thanks for this @mkerin it's a very good question!

In almost all cases where there is a lower BLAST hit with a valid start codon, the same CDS will be covered by the top hit (based on our tests). As a toy example, if a terL was 600 AA long, you could have a top BLAST hit from 200-600AA, and lower hit from 1-100AA, while 100-200AA has low homology to the dnaapler database and so returns no hit. In this case, the correct CDS would be found using the pyrodigal fallback method, as the correct CDS will have the most overlap with the best hit anyway (from 200-600AA). I anticipate this case will cover 90%+ of the cases where a lower hit begins with a valid start codon.

Chromosomes, phages and plasmids _should_ have only one of dnaA/repA/terL (of course there's exceptions to every rule because this is biology but as a general rule this holds).

So in the case where a replicon has more than 1 dnaA/terL/repA CDS in the genome, then presumably both should have BLAST hits (whether or not they start with a valid start codon). Because the top hit will be better (usually both in terms of length and homology), the pyrodigal fallback would choose that CDS with the best match regardless of start codon which I think is desired behaviour - to be honest, in this case it doesn't really matter which one is chosen.

The other edge case I can conceive of is where the original dnaA/terL/repA CDS is disrupted and is annotated as 2 different CDS - in fact that is what is occurring in the SAOMS1 phage test that you ran in the original comment in this thread (which I presume was a reason behind this question). This is a tricky edge case example hence why I included it in the tests.

In this case, the terL is disrupted by a mobile genetic element (HNH endonuclease). I've attached sample [pharokka](https://github.com/gbouras13/pharokka) output below from the .gff file, each row is a CDS. The start coordinate is the column after the 'CDS', the stop is the column after that. 'product' = the annotated function according to pharokka (I've bolded it also).

```
MW460250_1_subset	Pyrodigal_3.1.0	CDS	2580	2777	26.9	+	0	ID=ACPGARRA_CDS_0009;transl_table=11;phrog=5340;top_hit=No_MMseqs_PHROG_hit;locus_tag=ACPGARRA_CDS_0009;function=head and packaging;product=**terminase large subunit**
MW460250_1_subset	Pyrodigal_3.1.0	CDS	3071	4042	140.8	+	0	ID=ACPGARRA_CDS_0010;transl_table=11;phrog=559;top_hit=No_MMseqs_PHROG_hit;locus_tag=ACPGARRA_CDS_0010;function=DNA, RNA and nucleotide metabolism;product=**HNH endonuclease**
MW460250_1_subset	Pyrodigal_3.1.0	CDS	4183	5730	232.5	+	0	ID=ACPGARRA_CDS_0011;transl_table=11;phrog=675;top_hit=No_MMseqs_PHROG_hit;locus_tag=ACPGARRA_CDS_0011;function=head and packaging;product=**terminase large subunit**
MW
```

In this case, there are 2 CDS annotated as the terL - and in your attached BLAST output, the top hit is the second CDS beginning 4183 of length 540AA, and the second hit (starting with a valid start codon ) begins with 2580 length 65AA.

With the current logic, dnaapler will choose the second CDS using the fallback. I think this is actually the correct approach for this phage (knowing a bit (not much!) about phage biology suggests that the 4183 CDS is much more likely to be the 'real' truncated terL in this phage due to its length and the domains preserved).  Generally speaking if the disruption is early in the CDS, then the longer second hit (with no valid start codon in the BLAST hit) will likely be better.

In other cases where this happens, this may not be the case. If the disruption happens later in the original CDS, then the top hit (likely with a valid start codon)  will chosen be the longer/best hit of the 2 CDS.

To summarise, I think the fallback implemented is the best approach in most cases - but not perfect of course!

George
