+++ docs/example.md
@@ -51,7 +51,7 @@ The output should look like:
 2023-11-07 12:09:01.798 | INFO     | dnaapler.utils.util:end_dnaapler:159 - Elapsed time: 10.55 seconds
 ```
 
-In the results in the output directory, you will see that the `C333_reorientation_summary.tsv` file shows that `dnaapler` has identified the C333 genome to begin with coordinate
+In the results in the output directory, you will see that the `C333_reorientation_summary.tsv` file shows that `dnaapler` has identified the C333 genome to begin with coordinate 466140 on the reverse strand.
 
 A comparison of circular genomic maps of the C333 chromosome before and after `dnaapler` made with [Bakta v1.8.2](https://github.com/oschwengers/bakta) can be seen below. The genome begins at the top of the circle. Each thin line represents a CDS, with the positive stranded CDSs denoted in the outer ring and the negative stranded CDSs in the inner ring. The position of the chromosomal replication initiator gene is labelled as dnaA. The red and green ring denotes the GC content while the blue and yellow ring denotes the GC skew. 
 
