### 2023-11-13 gbouras13/dnaapler#48 @vinisalazar
LGTM @gbouras13.

I will close this issue for now but @mkerin or @gbouras13 please reopen it if you deem it necessary.
