### 2023-11-13 gbouras13/dnaapler#47 @vinisalazar
Thank you @gbouras13, I will now close this issue.
