### 2023-11-14 gbouras13/dnaapler#52 @mkerin
Thank you for the comprehensive response to this. I am happy to mark the issue as resolved.
