+++ paper/paper.bib
@@ -160,17 +160,22 @@
     eprint = {https://academic.oup.com/bioinformatics/article-pdf/39/7/btad409/50836034/btad409.pdf},
 }
 
-@article {Houtak:2023,
-	author = {Houtak, Ghais and Bouras, George and Nepal, Roshan and Shaghayegh, Gohar and Cooksley, Clare and  Psaltis, Alkis James and Wormald, Peter-John and Vreugde, Sarah},
-	title = {The Intra-Host Evolutionary Landscape And Pathoadaptation Of Persistent Staphylococcus aureus In Chronic Rhinosinusitis},
-	elocation-id = {2023.03.28.534496},
-	year = {2023},
-	doi = {10.1101/2023.03.28.534496},
-	publisher = {Cold Spring Harbor Laboratory},
-  URL = {https://www.biorxiv.org/content/early/2023/03/28/2023.03.28.534496},
-	eprint = {https://www.biorxiv.org/content/early/2023/03/28/2023.03.28.534496.full.pdf},
-	journal = {bioRxiv}
-}
+@article{Houtak:2023,
+   author = "Houtak, Ghais and Bouras, George and Nepal, Roshan and Shaghayegh, Gohar and Cooksley, Clare and Psaltis, Alkis James and Wormald, Peter-John and Vreugde, Sarah",
+   title = "The intra-host evolutionary landscape and pathoadaptation of persistent Staphylococcus aureus in chronic rhinosinusitis", 
+   journal= "Microbial Genomics",
+   year = "2023",
+   volume = "9",
+   number = "11",
+   pages = "",
+   doi = "https://doi.org/10.1099/mgen.0.001128",
+   url = "https://www.microbiologyresearch.org/content/journal/mgen/10.1099/mgen.0.001128",
+   publisher = "Microbiology Society",
+   issn = "2057-5858",
+   type = "Journal Article",
+   keywords = "S. aureus, Chronic Rhinosinusitis, Biofilm, Evolution, Plasmids",
+    eid = "001128",
+  }
 
 @article{Bouras2:2023,
     author = {Bouras, George and Nepal, Roshan and Houtak, Ghais and Psaltis, Alkis James and Wormald, Peter-John and Vreugde, Sarah},
