+++ paper/paper.md
@@ -19,7 +19,7 @@ authors:
     affiliation: 3
   - name: Michael J. Roach
     orcid: 0000-0003-1488-5148
-    affiliation: 3
+    affiliation: "3, 4"
 affiliations:
   - name: Adelaide Medical School, Faculty of Health and Medical Sciences, The University of Adelaide, Adelaide, South Australia 5005, Australia
     index: 1
@@ -27,7 +27,9 @@ affiliations:
     index: 2
   - name: Flinders Accelerator for Microbiome Exploration, College of Science and Engineering, Flinders University, Bedford Park, Adelaide, South Australia 5042, Australia
     index: 3
-date: 08 August 2023  
+  - name: Adelaide Centre for Epigenetics and South Australian Immunogenomics Cancer Institute, The University of Adelaide, Adelaide, South Australia 5005, Australia
+    index: 4
+date: 03 December 2023  
 bibliography: paper.bib
 ---
 
