### 2023-12-02 #5968 @gbouras13
Hi @fboehm ,

I have checked and confirm all authors and affiliations (adding a recent one for Michael Roach).

I have also updated one of the citations from bioRxiv to be the journal article (it had since been published)

I have checked the paper, it looks good to me.

The release version is v0.5.0.

The DOI of the Zenodo link is: https://doi.org/10.5281/zenodo.10251242

All the authors, orchids, affiliations and title match between the zenodo release and the paper.

Let me know if I need to fix anything!

George
