### 2023-12-20 #5968 @fboehm
@gbouras13 - it looks like two of the dois for your references have preceding text that needs to be removed. Would you please edit the bib file by deleting the "https://doi.org/" from the two doi entries?
