### 2023-12-20 #5968 @fboehm
IN other words, the doi field in the bib entry should be merely: 10.1099/mgen.0.001128 (for example)
