### 2023-12-20 #5968 @fboehm
Also, @gbouras13 - the software repo (https://github.com/gbouras13/dnaapler) and the software archive (https://doi.org/10.5281/zenodo.10251242) need to use the same license. It looks to me that the github repo uses MIT license, while the zenodo archive, I think, lists a creative commons license. Am I interpreting this accurately?

![image](https://github.com/openjournals/joss-reviews/assets/4155196/641018eb-3f98-4480-a22b-687f26a6ebdf)
