### 2023-12-20 #5968 @fboehm
I'm thinking that the zenodo page's "Rights" section should list MIT license.
