### 2023-12-20 #5968 @gbouras13
Hi @fboehm ,

Thanks for this - I have updated the zenodo page's rights to MIT and fixed those 2 spurious dois
