### 2023-12-27 #5968 @fboehm
lines 73 & 74: use lower case letters in continuous integration and continuous integration
@gbouras13 - can you fix this small issue? Otherwise, the text of the pdf looks good.
