### 2023-12-27 #5968 @fboehm
Also, the dois all resolve to the intended targets.
