### 2023-12-30 #5968 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
