### 2024-01-08 9e9afd8
add --db for all
