+++ src/dnaapler/utils/all.py
@@ -221,7 +221,6 @@ def all_process_blast_output_and_reorient(
             # so in these cases, go down the list and check there is a hit with a legitimate start codon
             # I anticipate this will be rare usually, the first will be it :)
             else:
-
                 # get gene
                 # set as dnaA by default
                 gene = "dnaA"
@@ -234,7 +233,9 @@ def all_process_blast_output_and_reorient(
                     gene = "terL"
 
                 # update the record description to contain 'rotated=True' akin to how unicycler does it
-                record.description = f"{record.description} rotated=True rotated_gene={gene}" 
+                record.description = (
+                    f"{record.description} rotated=True rotated_gene={gene}"
+                )
 
                 if filtered_df["qseq"][0][0] in ["M", "V", "L"] and (
                     filtered_df["sstart"][0] == 1
