+++ tests/test_overall.py
@@ -202,11 +202,13 @@ def test_all(tmp_dir):
     cmd = f"dnaapler all  -i {input_fasta} -o {tmp_dir} -t 1 -f"
     exec_command(cmd)
 
+
 def test_all_dnaa_repa(tmp_dir):
     """test all dnaa repa"""
     input_fasta: Path = f"{overall_test_data}/all_test.fasta"
     cmd = f"dnaapler all  -i {input_fasta} -o {tmp_dir} -t 1 -f --db dnaa,repa"
     exec_command(cmd)
+
 
 def test_all_dnaa_terl(tmp_dir):
     """test all dnaa terl"""
@@ -214,11 +216,13 @@ def test_all_dnaa_terl(tmp_dir):
     cmd = f"dnaapler all  -i {input_fasta} -o {tmp_dir} -t 1 -f --db dnaa,terl"
     exec_command(cmd)
 
+
 def test_all_repa_terl(tmp_dir):
     """test all repa terl"""
     input_fasta: Path = f"{overall_test_data}/all_reorient_and_no_reorient.fasta"
     cmd = f"dnaapler all  -i {input_fasta} -o {tmp_dir} -t 1 -f --db repa,terl"
     exec_command(cmd)
+
 
 def test_all_ignore(tmp_dir):
     """test all"""
