+++ src/dnaapler/utils/bulk.py
@@ -52,11 +52,11 @@ def run_bulk_blast(
     elif gene == "all":
         db_name = "all_db"
     elif gene == "dnaA,repA":
-        gene = "dnaA_repA_db"
+        db_name = "dnaA_repA_db"
     elif gene == "dnaA,terL":
-        gene = "dnaA_terL_db"
+        db_name = "dnaA_terL_db"
     elif gene == "repA,terL":
-        gene = "repA_terL_db"
+        db_name = "repA_terL_db"
 
     # for chromosome, plasmid or phage or all
     # runs blast
