### 2024-01-09 cea0db9
bump version, also make release easier with ci
